create package sercom
* use package sercom
